Fixed rich presence and updated csproj
fixed rich presence by adding a timer on the update method, also added boss reference. Updated csproj.

+++ Globals/ROIGlobalNPC.cs
@@ -108,7 +108,7 @@ namespace ROI.Globals
             }
         }
 
-        private void RewardVoidTier(int tier)
+        private void RewardVoidTier(byte tier)
         {
             foreach (var player in Main.player)
             {
