+++ Players/Void/ROIVoidPlayer.cs
@@ -25,14 +25,13 @@ namespace ROI.Players
 			int simulatedAmount = Math.Min(MaxVoidAffinity - VoidAffinityAmount, voidAffinity);
 			if (!simulate)
 			{
-				_voidAffinityAmount += simulatedAmount;
+				_voidAffinityAmount += (short)simulatedAmount;
 			}
 			return simulatedAmount;
 		}
 
 		public void DamageVoidHeart(ref int damage)
-		{
-
+        {
 			if (VoidHeartHP >= 0)
 			{
 			    CombatText.NewText(
